+++ Thunderstore/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2026-05-07
+
+### Added
+
+- LethalConstellations compatibility with generated per-constellation deadline config [ `ConfigurableQuota_Constellations.cfg` ]
+- Per-constellation deadline mode - `UseGlobal`, `Fixed` or `Random`
+
 ## [1.0.2] - 2026-05-07
 
 ### Fixed